fix(global): :sparkles: defaultAppBarContent added and changes made to CoreAppBar
defaultAppBarContent added and changes made to CoreAppBar

ref: #142
